[level 0] Title: 배열의 평균값, Time: 0.04 ms, Memory: 83.2 MB -BaekjoonHub

+++ 프로그래머스/lv0/120817. 배열의 평균값/배열의 평균값.java
@@ -1,6 +1,7 @@
 class Solution {
     public double solution(int[] numbers) {
-        float sum = 0;
+        //float vs double
+        double sum = 0;
         for(int i=0; i< numbers.length; i++){
            sum = sum + numbers[i];
         }
